requirements.txt
streamlit
pandas
numpy
shap
scikit-learn
joblib
mysql-connector-python
matplotlib